python/requirements.txt
Cerberus==0.8.1
Eve==0.5.3
Events==0.2.1
Flask==0.10.1
Flask-PyMongo==0.3.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.10.4
itsdangerous==0.24
py2neo==2.0.6
pymongo==2.8
simplejson==3.6.5
virtualenv==1.11.6
wsgiref==0.1.2
flask-cors
PyYAML
pymongo
